+++ app/src/main/kotlin/com/simplemobiletools/camera/Constants.kt
@@ -11,6 +11,7 @@ val SOUND = "sound"
 val FOCUS_BEFORE_CAPTURE = "focus_before_capture"
 val VOLUME_BUTTONS_AS_SHUTTER = "volume_buttons_as_shutter"
 val TURN_FLASH_OFF_AT_STARTUP = "turn_flash_off_at_startup"
+val FLIP_PHOTOS = "flip_photos"
 val LAST_USED_CAMERA = "last_used_camera"
 val FLASHLIGHT_STATE = "flashlight_state"
 val BACK_PHOTO_RESOLUTION_INDEX = "back_photo_resolution_index"
@@ -18,6 +19,9 @@ val BACK_VIDEO_RESOLUTION_INDEX = "back_video_resolution_index"
 val FRONT_PHOTO_RESOLUTION_INDEX = "front_photo_resolution_index"
 val FRONT_VIDEO_RESOLUTION_INDEX = "front_video_resolution_index"
 val PHOTO_PREVIEW_HINT_SHOWN = "photo_preview_hint_shown"
+val KEEP_SETTINGS_VISIBLE = "keep_settings_visible"
+val ALWAYS_OPEN_BACK_CAMERA = "always_open_back_camera"
+val SAVE_PHOTO_METADATA = "save_photo_metadata"
 
 val FLASH_OFF = 0
 val FLASH_ON = 1