+++ app/src/main/kotlin/com/simplemobiletools/camera/Preview.kt
@@ -9,7 +9,6 @@ import android.net.Uri
 import android.os.Build
 import android.os.Environment
 import android.os.Handler
-import android.util.Log
 import android.view.ScaleGestureDetector
 import android.view.SurfaceHolder
 import android.view.SurfaceView
@@ -24,7 +23,6 @@ import java.util.*
 
 class Preview : ViewGroup, SurfaceHolder.Callback, MediaScannerConnection.OnScanCompletedListener {
     var mCamera: Camera? = null
-    private val TAG = Preview::class.java.simpleName
     private val FOCUS_AREA_SIZE = 100
     private val PHOTO_PREVIEW_LENGTH = 500L
     private val REFOCUS_PERIOD = 3000L
@@ -39,9 +37,8 @@ class Preview : ViewGroup, SurfaceHolder.Callback, MediaScannerConnection.OnScan
     private var mPreviewSize: Camera.Size? = null
     private var mParameters: Camera.Parameters? = null
     private var mRecorder: MediaRecorder? = null
-    private var mTargetUri: Uri? = null
     private var mScaleGestureDetector: ScaleGestureDetector? = null
-    private var mZoomRatios: List<Int>? = null
+    private var mZoomRatios = ArrayList<Int>()
 
     private var mCurrVideoPath = ""
     private var mCanTakePicture = false
@@ -62,6 +59,8 @@ class Preview : ViewGroup, SurfaceHolder.Callback, MediaScannerConnection.OnScan
     private var autoFocusHandler = Handler()
 
     var isWaitingForTakePictureCallback = false
+    var mTargetUri: Uri? = null
+    var isImageCaptureIntent = false
 
     constructor(context: Context) : super(context)
 
@@ -94,7 +93,7 @@ class Preview : ViewGroup, SurfaceHolder.Callback, MediaScannerConnection.OnScan
             if (mIsPreviewShown) {
                 resumePreview()
             } else {
-                if (!mWasZooming)
+                if (!mWasZooming && !mIsPreviewShown)
                     focusArea(false)
 
                 mWasZooming = false
@@ -118,8 +117,7 @@ class Preview : ViewGroup, SurfaceHolder.Callback, MediaScannerConnection.OnScan
             newCamera = Camera.open(cameraId)
             mCallback.setIsCameraAvailable(true)
         } catch (e: Exception) {
-            mActivity.toast(R.string.camera_open_error)
-            Log.e(TAG, "setCamera open " + e.message)
+            mActivity.showErrorToast(e)
             mCallback.setIsCameraAvailable(false)
             return false
         }
@@ -143,7 +141,10 @@ class Preview : ViewGroup, SurfaceHolder.Callback, MediaScannerConnection.OnScan
 
         mParameters = mCamera!!.parameters
         mMaxZoom = mParameters!!.maxZoom
-        mZoomRatios = mParameters!!.zoomRatios
+
+        if (mParameters!!.isZoomSupported)
+            mZoomRatios = mParameters!!.zoomRatios as ArrayList<Int>
+
         mSupportedPreviewSizes = mParameters!!.supportedPreviewSizes.sortedByDescending { it.width * it.height }
         refreshPreview()
 
@@ -158,7 +159,7 @@ class Preview : ViewGroup, SurfaceHolder.Callback, MediaScannerConnection.OnScan
             try {
                 mCamera!!.setPreviewDisplay(mSurfaceHolder)
             } catch (e: IOException) {
-                Log.e(TAG, "setCamera setPreviewDisplay " + e.message)
+                mActivity.showErrorToast(e)
                 return false
             }
         }
@@ -214,33 +215,29 @@ class Preview : ViewGroup, SurfaceHolder.Callback, MediaScannerConnection.OnScan
         return null
     }
 
-    fun setTargetUri(uri: Uri) {
-        mTargetUri = uri
-    }
-
     private fun initGestureDetector() {
         mScaleGestureDetector = ScaleGestureDetector(mActivity, object : ScaleGestureDetector.SimpleOnScaleGestureListener() {
             override fun onScale(detector: ScaleGestureDetector): Boolean {
                 val zoomFactor = mParameters!!.zoom
-                var zoomRatio = mZoomRatios!![zoomFactor] / 100f
+                var zoomRatio = mZoomRatios[zoomFactor] / 100f
                 zoomRatio *= detector.scaleFactor
 
                 var newZoomFactor = zoomFactor
                 if (zoomRatio <= 1f) {
                     newZoomFactor = 0
-                } else if (zoomRatio >= mZoomRatios!![mMaxZoom] / 100f) {
+                } else if (zoomRatio >= mZoomRatios[mMaxZoom] / 100f) {
                     newZoomFactor = mMaxZoom
                 } else {
                     if (detector.scaleFactor > 1f) {
-                        for (i in zoomFactor..mZoomRatios!!.size - 1) {
-                            if (mZoomRatios!![i] / 100.0f >= zoomRatio) {
+                        for (i in zoomFactor until mZoomRatios.size) {
+                            if (mZoomRatios[i] / 100.0f >= zoomRatio) {
                                 newZoomFactor = i
                                 break
                             }
                         }
                     } else {
                         for (i in zoomFactor downTo 0) {
-                            if (mZoomRatios!![i] / 100.0f <= zoomRatio) {
+                            if (mZoomRatios[i] / 100.0f <= zoomRatio) {
                                 newZoomFactor = i
                                 break
                             }
@@ -285,15 +282,19 @@ class Preview : ViewGroup, SurfaceHolder.Callback, MediaScannerConnection.OnScan
             mRotationAtCapture = MainActivity.mLastHandledOrientation
             mCamera!!.parameters = mParameters
             isWaitingForTakePictureCallback = true
-            mCamera!!.takePicture(null, null, takePictureCallback)
+            mIsPreviewShown = true
+            try {
+                mCamera!!.takePicture(null, null, takePictureCallback)
 
-            if (config.isSoundEnabled) {
-                val audioManager = context.getSystemService(Context.AUDIO_SERVICE) as AudioManager
-                val volume = audioManager.getStreamVolume(AudioManager.STREAM_SYSTEM)
-                if (volume != 0) {
-                    val mp = MediaPlayer.create(context, Uri.parse("file:///system/media/audio/ui/camera_click.ogg"))
-                    mp?.start()
+                if (config.isSoundEnabled) {
+                    val audioManager = context.getSystemService(Context.AUDIO_SERVICE) as AudioManager
+                    val volume = audioManager.getStreamVolume(AudioManager.STREAM_SYSTEM)
+                    if (volume != 0) {
+                        val mp = MediaPlayer.create(context, Uri.parse("file:///system/media/audio/ui/camera_click.ogg"))
+                        mp?.start()
+                    }
                 }
+            } catch (ignored: Exception) {
             }
         }
         mCanTakePicture = false
@@ -302,32 +303,54 @@ class Preview : ViewGroup, SurfaceHolder.Callback, MediaScannerConnection.OnScan
 
     private val takePictureCallback = Camera.PictureCallback { data, cam ->
         isWaitingForTakePictureCallback = false
+        if (!isImageCaptureIntent) {
+            handlePreview()
+        }
+
+        if (isImageCaptureIntent) {
+            if (mTargetUri != null) {
+                storePhoto(data)
+            } else {
+                mActivity.finishActivity()
+            }
+        } else {
+            storePhoto(data)
+        }
+    }
+
+    private fun storePhoto(data: ByteArray) {
+        PhotoProcessor(mActivity, mTargetUri, mCurrCameraId, mRotationAtCapture).execute(data)
+    }
+
+    private fun handlePreview() {
         if (config.isShowPreviewEnabled) {
-            mIsPreviewShown = true
             if (!config.wasPhotoPreviewHintShown) {
                 mActivity.toast(R.string.click_to_resume_preview)
                 config.wasPhotoPreviewHintShown = true
             }
         } else {
             Handler().postDelayed({
+                mIsPreviewShown = false
                 resumePreview()
             }, PHOTO_PREVIEW_LENGTH)
         }
-
-        PhotoProcessor(mActivity, mTargetUri, mCurrCameraId, mRotationAtCapture).execute(data)
     }
 
     private fun resumePreview() {
         mIsPreviewShown = false
         mActivity.toggleBottomButtons(false)
-        mCamera?.startPreview()
+        try {
+            mCamera?.startPreview()
+        } catch (ignored: Exception) {
+        }
         mCanTakePicture = true
         focusArea(false, false)
     }
 
     private fun focusArea(takePictureAfter: Boolean, showFocusRect: Boolean = true) {
-        if (mCamera == null || (mIsFocusingBeforeCapture && !takePictureAfter))
+        if (mCamera == null || (mIsFocusingBeforeCapture && !takePictureAfter)) {
             return
+        }
 
         if (takePictureAfter)
             mIsFocusingBeforeCapture = true
@@ -347,9 +370,13 @@ class Preview : ViewGroup, SurfaceHolder.Callback, MediaScannerConnection.OnScan
                 mCallback.drawFocusRect(mLastClickX, mLastClickY)
         }
 
-        mCamera!!.parameters = mParameters
         try {
+            mCamera!!.parameters = mParameters
             mCamera!!.autoFocus { success, camera ->
+                if (camera == null || mCamera == null) {
+                    return@autoFocus
+                }
+
                 camera.cancelAutoFocus()
                 val focusModes = mParameters!!.supportedFocusModes
                 if (focusModes.contains(Camera.Parameters.FOCUS_MODE_CONTINUOUS_PICTURE))
@@ -363,8 +390,7 @@ class Preview : ViewGroup, SurfaceHolder.Callback, MediaScannerConnection.OnScan
                     rescheduleAutofocus()
                 }
             }
-        } catch (ignored: RuntimeException) {
-
+        } catch (ignored: Exception) {
         }
     }
 
@@ -419,7 +445,7 @@ class Preview : ViewGroup, SurfaceHolder.Callback, MediaScannerConnection.OnScan
             if (mSwitchToVideoAsap)
                 initRecorder()
         } catch (e: IOException) {
-            Log.e(TAG, "surfaceCreated IOException " + e.message)
+            mActivity.showErrorToast(e)
         }
     }
 
@@ -447,7 +473,8 @@ class Preview : ViewGroup, SurfaceHolder.Callback, MediaScannerConnection.OnScan
             mCamera!!.parameters = mParameters
             try {
                 mCamera!!.startPreview()
-            } catch (ignored: RuntimeException) {
+            } catch (e: RuntimeException) {
+                mActivity.showErrorToast(e)
             }
         }
     }
@@ -463,9 +490,7 @@ class Preview : ViewGroup, SurfaceHolder.Callback, MediaScannerConnection.OnScan
         }
     }
 
-    override fun onLayout(changed: Boolean, l: Int, t: Int, r: Int, b: Int) {
-
-    }
+    override fun onLayout(changed: Boolean, l: Int, t: Int, r: Int, b: Int) {}
 
     private fun getOptimalPreviewSize(sizes: List<Camera.Size>, width: Int, height: Int): Camera.Size {
         var result: Camera.Size? = null
@@ -493,12 +518,12 @@ class Preview : ViewGroup, SurfaceHolder.Callback, MediaScannerConnection.OnScan
         if (mSupportedPreviewSizes != null) {
             // for simplicity lets assume that most displays are 16:9 and the remaining ones are 4:3
             // always set 16:9 for videos as many devices support 4:3 only in low quality
-            if (mIsSixteenToNine || mIsVideoMode) {
-                mPreviewSize = getOptimalPreviewSize(mSupportedPreviewSizes!!, mScreenSize.y, mScreenSize.x)
+            mPreviewSize = if (mIsSixteenToNine || mIsVideoMode) {
+                getOptimalPreviewSize(mSupportedPreviewSizes!!, mScreenSize.y, mScreenSize.x)
             } else {
                 val newRatioHeight = (mScreenSize.x * (4.toDouble() / 3)).toInt()
                 setMeasuredDimension(mScreenSize.x, newRatioHeight)
-                mPreviewSize = getOptimalPreviewSize(mSupportedPreviewSizes!!, newRatioHeight, mScreenSize.x)
+                getOptimalPreviewSize(mSupportedPreviewSizes!!, newRatioHeight, mScreenSize.x)
             }
             val lp = mSurfaceView.layoutParams
 
@@ -630,8 +655,7 @@ class Preview : ViewGroup, SurfaceHolder.Callback, MediaScannerConnection.OnScan
     }
 
     private fun setupFailed(e: Exception) {
-        mActivity.toast(R.string.video_setup_error)
-        Log.e(TAG, "initRecorder " + e.message)
+        mActivity.showErrorToast(e)
         releaseCamera()
     }
 
@@ -666,8 +690,7 @@ class Preview : ViewGroup, SurfaceHolder.Callback, MediaScannerConnection.OnScan
             toggleShutterSound(false)
             mIsRecording = true
         } catch (e: Exception) {
-            mActivity.toast(R.string.video_setup_error)
-            Log.e(TAG, "toggleRecording " + e.message)
+            mActivity.showErrorToast(e)
             releaseCamera()
         }
     }
@@ -679,10 +702,9 @@ class Preview : ViewGroup, SurfaceHolder.Callback, MediaScannerConnection.OnScan
                 mRecorder!!.stop()
                 mActivity.scanPath(mCurrVideoPath) {}
             } catch (e: RuntimeException) {
+                mActivity.showErrorToast(e)
                 toggleShutterSound(false)
                 File(mCurrVideoPath).delete()
-                mActivity.toast(R.string.video_saving_error)
-                Log.e(TAG, "stopRecording " + e.message)
                 mRecorder = null
                 mIsRecording = false
                 releaseCamera()