+++ app/src/main/kotlin/com/simplemobiletools/camera/extensions/size.kt
@@ -13,9 +13,23 @@ fun Camera.Size.isSixteenToNine(): Boolean {
     return diff < RATIO_TOLERANCE
 }
 
+fun Camera.Size.isFiveToThree(): Boolean {
+    val selectedRatio = Math.abs(width / height.toFloat())
+    val checkedRatio = 5 / 3.toFloat()
+    val diff = Math.abs(selectedRatio - checkedRatio)
+    return diff < RATIO_TOLERANCE
+}
+
 fun Camera.Size.isFourToThree(): Boolean {
     val selectedRatio = Math.abs(width / height.toFloat())
     val checkedRatio = 4 / 3.toFloat()
+    val diff = Math.abs(selectedRatio - checkedRatio)
+    return diff < RATIO_TOLERANCE
+}
+
+fun Camera.Size.isThreeToFour(): Boolean {
+    val selectedRatio = Math.abs(width / height.toFloat())
+    val checkedRatio = 3 / 4.toFloat()
     val diff = Math.abs(selectedRatio - checkedRatio)
     return diff < RATIO_TOLERANCE
 }
@@ -34,16 +48,18 @@ fun Camera.Size.isSixToFive(): Boolean {
     return diff < RATIO_TOLERANCE
 }
 
-fun Camera.Size.getAspectRatio(context: Context): String {
-    return if (isSixteenToNine()) {
-        "16:9"
-    } else if (isFourToThree()) {
-        "4:3"
-    } else if (isThreeToTwo()) {
-        "3:2"
-    } else if (isSixToFive()) {
-        "6:5"
-    } else {
-        context.resources.getString(R.string.other)
-    }
+fun Camera.Size.isOneNineToOne() = Math.abs(1.9 - (width / height.toFloat())) < RATIO_TOLERANCE
+
+fun Camera.Size.isSquare() = width == height
+
+fun Camera.Size.getAspectRatio(context: Context) = when {
+    isSixteenToNine() -> "16:9"
+    isFiveToThree() -> "5:3"
+    isFourToThree() -> "4:3"
+    isThreeToFour() -> "3:4"
+    isThreeToTwo() -> "3:2"
+    isSixToFive() -> "6:5"
+    isOneNineToOne() -> "1.9:1"
+    isSquare() -> "1:1"
+    else -> context.resources.getString(R.string.other)
 }